Move FavoriteAnimation to core module
