+++ uicomponent-compose/core/src/main/res/drawable/ic_baseline_favorite_24.xml
