+++ uicomponent-compose/core/src/main/res/drawable/ic_baseline_favorite_border_24.xml
